chore: moved to latest version of use-agent; renamed imports
