+++ ui/apps/dashboard/package.json
@@ -29,8 +29,8 @@
     "@graphql-typed-document-node/core": "^3.2.0",
     "@headlessui/react": "^1.7.17",
     "@headlessui/tailwindcss": "^0.1.2",
-    "@inngest/agent-kit": "0.11.0-alpha-20250924000141-d3d24028fd576193e2adfa963a69864a2fc46d94",
-    "@inngest/use-agents": "0.2.0-alpha-20250924000141-d3d24028fd576193e2adfa963a69864a2fc46d94",
+    "@inngest/agent-kit": "0.11.0",
+    "@inngest/use-agent": "0.2.0",
     "@inngest/components": "workspace:*",
     "@inngest/realtime": "^0.4.0",
     "@launchdarkly/node-server-sdk": "^9.0.3",
