+++ ui/apps/dashboard/src/components/Insights/InsightsChat/InsightsChat.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { type AgentStatus } from '@inngest/use-agents';
+import { type AgentStatus } from '@inngest/use-agent';
 
 import { useAllEventTypes } from '@/components/EventTypes/useEventTypes';
 import { useInsightsStateMachineContext } from '@/components/Insights/InsightsStateMachineContext/InsightsStateMachineContext';
